Fix errors when trying to run current Windows container images that do not have a D drive available for mapping onto.
Instead of trying to map specific GitHub dirs from D to C, this maps the entirety of the GitHub runner directory (i.e. D:/a) to the C drive. To allow for that all environment variables that refer to the D: drive are rewritten to refer to the C: drive instead.

An additional change in this is to switch to the Windows 2022 OS and container image for testing. As the 2019 OS was removed. There's also a change to allow testing on fix/ branches to allow for forks to create local test branches ahead of creating PRs.